httpserver: log request response pairs to the handler
Log the request-response pairs to the handler, so it can be verified in the
test code.

This is very similar to the log attribute defined on the server which
contains all requests and responses, the difference is that in this case the
handler object itself has a log about it so it could be verified that the
request was handled by the given handler.

This commit also updates the API: methods starting with "respond_with_" will
return the handler itself so the calls attribute can be accessed easily.